Migración a Base de Datos No Relacional con MongoDB/requirements.txt
pymongo==4.3.3
python-dotenv==1.0.0
colorama==0.4.6
tabulate==0.9.0